# feat: replace placeholder benchmarks with meaningful MCP tool performance tests