… tests/benchmarks/sample.bench.ts