## Summary

Replace placeholder benchmarks with meaningful MCP tool performance tests using the real production database.

## Motivation

The existing benchmark suite consisted of:
- `sample.bench.ts` - Generic JavaScript benchmarks (array sorting, string concatenation) that are irrelevant to n8n-mcp
- `database-queries.bench.ts` - Database layer benchmarks using **mock data** (500 artificial nodes)

Neither of these benchmarks measured what actually matters: the performance that AI assistants experience when calling MCP tools in real-world scenarios.

## Changes

### Deleted
- ❌ `tests/benchmarks/sample.bench.ts` - Removed irrelevant generic JS benchmarks

### Added
- ✅ `tests/benchmarks/mcp-tools.bench.ts` - **8 real-world performance benchmarks** using production database (525+ nodes):

  **1. search_nodes (2 benchmarks)**
  - Common query: "http"
  - AI agent query: "slack send message"
  - Target: <20ms
  
  **2. get_node_essentials (2 benchmarks)**
  - HTTP Request node property filtering
  - Slack node property filtering
  - Target: <10ms
  
  **3. list_nodes (2 benchmarks)**
  - First 50 nodes
  - AI tools only
  - Target: <15ms
  
  **4. validate_node_operation (2 benchmarks)**
  - HTTP Request minimal validation
  - HTTP Request with parameters
  - Target: <15ms

### Modified
- 📝 `tests/benchmarks/database-queries.bench.ts` - Added comment clarifying it uses mock data
- 📝 `tests/benchmarks/index.ts` - Export new benchmark suite

## Why This Matters

These benchmarks measure the **actual performance AI assistants experience** when calling MCP tools, which is the most meaningful performance metric for n8n-mcp. They use the real production database with 525+ nodes, reflecting real-world usage patterns.

## Test Plan

- ✅ TypeScript compilation passes
- ✅ All benchmarks execute successfully
- ✅ Benchmark configuration validated (iterations, warmup, timing)

## Next Steps

These benchmarks will run in CI and provide ongoing performance tracking for MCP tool operations.

🤖 Generated with [Claude Code](https://claude.com/claude-code)